# fix: Slow DB query for advocates map