--- a/src/bin/benchmarks/benchmarkTotalAdvocatesPerState.ts
+++ b/src/bin/benchmarks/benchmarkTotalAdvocatesPerState.ts
@@ -54,6 +54,36 @@ async function benchmarkAdvocatesPerState() {
   // console.log('Query Result:', result)
 }
 
+// New query with EXPLAIN
+async function getTotalAdvocatesByStateExplain(stateCode: USStateCode) {
+  return prismaClient.$queryRaw`
+    EXPLAIN
+    SELECT
+      address.administrative_area_level_1 AS state,
+      COUNT(user.id) AS totalAdvocates
+    FROM address
+    JOIN user ON user.address_id = address.id
+    WHERE address.administrative_area_level_1 = ${stateCode}
+    AND address.country_code = 'US'
+    GROUP BY address.administrative_area_level_1;
+  `
+}
+
+// New query with EXPLAIN ANALYZE
+async function getTotalAdvocatesByStateExplainAnalyze(stateCode: USStateCode) {
+  return prismaClient.$queryRaw`
+    EXPLAIN ANALYZE
+    SELECT
+      address.administrative_area_level_1 AS state,
+      COUNT(user.id) AS totalAdvocates
+    FROM address
+    JOIN user ON user.address_id = address.id
+    WHERE address.administrative_area_level_1 = ${stateCode}
+    AND address.country_code = 'US'
+    GROUP BY address.administrative_area_level_1;
+  `
+}
+
 async function queryStatePerformance(stateCode: USStateCode): Promise<{
   state: USStateCode
   duration: number
@@ -65,7 +95,7 @@ async function queryStatePerformance(stateCode: USStateCode): Promise<{
   return { state: stateCode, duration: end - start, result }
 }
 
-async function benchmarkAdvocatesPerStateByState() {
+async function benchmarkAdvocatesByState() {
   const stateCodes = Object.keys(US_STATE_CODE_TO_DISTRICT_COUNT_MAP) as USStateCode[]
   const statePromises = stateCodes.map(queryStatePerformance)
 
@@ -116,7 +146,16 @@ async function benchmarkAdvocatesPerStateByState() {
 
 async function main() {
   await benchmarkAdvocatesPerState()
-  await benchmarkAdvocatesPerStateByState()
+
+  console.log(`\n--- Running EXPLAIN for getTotalAdvocatesByState (CA) ---`)
+  const byStateExplain = await getTotalAdvocatesByStateExplain('CA')
+  console.log(byStateExplain)
+
+  console.log(`\n--- Running EXPLAIN ANALYZE for getTotalAdvocatesByState (CA) ---`)
+  const byStateExplainAnalyze = await getTotalAdvocatesByStateExplainAnalyze('CA')
+  console.log(byStateExplainAnalyze)
+
+  await benchmarkAdvocatesByState()
 }
 
 void runBin(main)