--- a/src/data/aggregations/getTotalAdvocatesPerState.ts
+++ b/src/data/aggregations/getTotalAdvocatesPerState.ts
@@ -2,21 +2,23 @@ import 'server-only'
 
 import { prismaClient } from '@/utils/server/prismaClient'
 import { NEXT_PUBLIC_ENVIRONMENT } from '@/utils/shared/sharedEnv'
+import { USStateCode } from '@/utils/shared/stateMappings/usStateUtils'
+import { DEFAULT_SUPPORTED_COUNTRY_CODE } from '@/utils/shared/supportedCountries'
 
 type TotalAdvocatesPerStateQuery = {
   state: string
   totalAdvocates: bigint
 }[]
 
-const fetchAllFromPrisma = async () => {
+const fetchAllFromPrismaByState = async (stateCode: USStateCode) => {
   return prismaClient.$queryRaw<TotalAdvocatesPerStateQuery>`
     SELECT
       address.administrative_area_level_1 AS state,
       COUNT(user.id) AS totalAdvocates
     FROM address
     JOIN user ON user.address_id = address.id
-    WHERE address.country_code = 'US'
-    AND address.administrative_area_level_1 != ''
+    WHERE address.administrative_area_level_1 = ${stateCode}
+    AND address.country_code = ${DEFAULT_SUPPORTED_COUNTRY_CODE}
     GROUP BY address.administrative_area_level_1;
   `
 }
@@ -35,8 +37,8 @@ const parseTotalAdvocatesPerState = (totalAdvocatesPerState: TotalAdvocatesPerSt
     .filter(({ state }) => state.length <= 2)
 }
 
-export const getTotalAdvocatesPerState = async () => {
-  const rawTotalAdvocatesPerState = await fetchAllFromPrisma()
+export const getTotalAdvocatesByState = async (stateCode: USStateCode) => {
+  const rawTotalAdvocatesPerState = await fetchAllFromPrismaByState(stateCode)
   const totalAdvocatesPerState = parseTotalAdvocatesPerState(rawTotalAdvocatesPerState)
 
   return totalAdvocatesPerState