--- a/src/data/pageSpecific/getAdvocatesMapData.ts
+++ b/src/data/pageSpecific/getAdvocatesMapData.ts
@@ -1,9 +1,21 @@
 import 'server-only'
 
-import { getTotalAdvocatesPerState } from '@/data/aggregations/getTotalAdvocatesPerState'
+import pRetry from 'p-retry'
+
+import { getTotalAdvocatesByState } from '@/data/aggregations/getTotalAdvocatesPerState'
+import {
+  US_STATE_CODE_TO_DISPLAY_NAME_MAP,
+  USStateCode,
+} from '@/utils/shared/stateMappings/usStateUtils'
 
 export async function getAdvocatesMapData() {
-  const [totalAdvocatesPerState] = await Promise.all([getTotalAdvocatesPerState()])
+  const stateCodes = Object.keys(US_STATE_CODE_TO_DISPLAY_NAME_MAP) as USStateCode[]
+  const advocatesPerStatePromises = stateCodes.map(stateCode =>
+    pRetry(async () => await getTotalAdvocatesByState(stateCode)),
+  )
+
+  const results = await Promise.all(advocatesPerStatePromises)
+  const totalAdvocatesPerState = results.flat()
 
   return {
     advocatesMapData: {
@@ -11,5 +23,4 @@ export async function getAdvocatesMapData() {
     },
   }
 }
-
 export type GetAdvocatesMapDataResponse = Awaited<ReturnType<typeof getAdvocatesMapData>>