closes #2292 

This PR fixes a slow database query on the advocates map page, improving performance by over 3x.

**Changes:**
- Replaced a single, slow query with parallel, state-by-state queries.
- Added a new rule for diagnosing and fixing slow database queries.

**Performance Improvement:**

The original query, which fetched data for all states at once, was inefficient. By splitting the logic into smaller, concurrent queries (one for each state), we allow the database to make much more effective use of the existing `@index(administrative_area_level_1, country_code)` index. Each smaller query is highly selective on `administrative_area_level_1` (the leading column of the index), allowing the database to quickly find the relevant data for each state. Running these highly efficient queries in parallel results in a total wall-clock time that is over 3x faster than the original approach.

### **Benchmarks:**
<details>
  <summary><strong>Original Query</strong></summary>
  
  ```
--- Running EXPLAIN for getTotalAdvocatesPerState ---
[
  {
    id: 1n,
    select_type: 'SIMPLE',
    table: 'address',
    partitions: null,
    type: 'range',
    possible_keys: 'PRIMARY,address_administrative_area_level_1_country_code_idx
',
    key: 'address_administrative_area_level_1_country_code_idx',
    key_len: '766',
    ref: null,
    rows: 918458n,
    filtered: 10,
    Extra: 'Using where; Using index'
  },
  {
    id: 1n,
    select_type: 'SIMPLE',
    table: 'user',
    partitions: null,
    type: 'ref',
    possible_keys: 'user_address_id_idx',
    key: 'user_address_id_idx',
    key_len: '767',
    ref: 'swc-web.address.id',
    rows: 1n,
    filtered: 100,
    Extra: 'Using index'
  }
]

--- Running EXPLAIN ANALYZE for getTotalAdvocatesPerState ---
[
  {
    EXPLAIN: '-> Group aggregate: count(`user`.id)  (cost=292879 rows=4817) (act
ual time=1.63..7004 rows=375
 loops=1)\n' +
                                  '    -> Nested loop inner join  (cost=280522 r
ows=123577) (actual time=0.0629..6831 rows=1.88e+6 loops
=1)\n' +
                                  "        -> Filter: ((address.country_code = '
US') and (address.administrative_area_level_1 <> ''))  (
cost=241095 rows=91846) (actual time=0.0477..787 rows=1.76e+6 loops=1)\n" +
                                  "            -> Covering index range scan on a
ddress using address_administrative_area_level_1_country
_code_idx over (administrative_area_level_1 < '') OR ('' < administrative_area_l
evel_1)  (cost=241095 rows=918458) (actual time=0.0433..530 rows=1.86e+6 loops=1
)\n" +                                                        '        -> Coveri
ng index lookup on user using user_address_id_idx (address_id=address.id)  (cost
=0.2
95 rows=1.35) (actual time=0.00288..0.00327 rows=1.07 loops=1.76e+6)\n'
                              }
]

--- Benchmarking getTotalAdvocatesPerState ---
Query executed in: 6773.68ms
```
  
 </details>


<details>
  <summary><strong>New Query</strong></summary>

```
--- Running EXPLAIN for getTotalAdvocatesPerState ---
[
  {
    id: 1n,
    select_type: 'SIMPLE',
    table: 'address',
    partitions: null,
    type: 'index',
    possible_keys: 'PRIMARY,address_administrative_area_level_1_country_code_idx',
    key: 'address_administrative_area_level_1_country_code_idx',
    key_len: '1532',
    ref: null,
    rows: 1n,
    filtered: 100,
    Extra: 'Using where; Using index'
  },
  {
    id: 1n,
    select_type: 'SIMPLE',
    table: 'user',
    partitions: null,
    type: 'ref',
    possible_keys: 'user_address_id_idx',
    key: 'user_address_id_idx',
    key_len: '767',
    ref: 'swc-web.address.id',
    rows: 1n,
    filtered: 100,
    Extra: 'Using index'
  }
]

--- Running EXPLAIN ANALYZE for getTotalAdvocatesPerState ---
[
  {
    EXPLAIN: '-> Group aggregate: count(`user`.id)  (cost=0.8 rows=1) (actual time=0.0169..0.0169 rows=0 loops=1)\n' +
      '    -> Nested loop inner join  (cost=0.7 rows=1) (actual time=0.0161..0.0161 rows=0 loops=1)\n' +
      "        -> Filter: ((address.administrative_area_level_1 <> '') and (address.country_code = 'US'))  (cost=0.35 rows=1) (actual time=0.0144..0.0144 rows=0 loops=1)\n" +
      '            -> Covering index scan on address using address_administrative_area_level_1_country_code_idx  (cost=0.35 rows=1) (actual time=0.0138..0.0138 rows=0 loops=1)\n' +
      '        -> Covering index lookup on user using user_address_id_idx (address_id=address.id)  (cost=0.35 rows=1) (never executed)\n'
  }
]

--- Running EXPLAIN for getTotalAdvocatesByState (CA) ---
[
  {
    id: 1n,
    select_type: 'SIMPLE',
    table: 'address',
    partitions: null,
    type: 'ref',
    possible_keys: 'PRIMARY,address_administrative_area_level_1_country_code_idx',
    key: 'address_administrative_area_level_1_country_code_idx',
    key_len: '1532',
    ref: 'const,const',
    rows: 1n,
    filtered: 100,
    Extra: 'Using index'
  },
  {
    id: 1n,
    select_type: 'SIMPLE',
    table: 'user',
    partitions: null,
    type: 'ref',
    possible_keys: 'user_address_id_idx',
    key: 'user_address_id_idx',
    key_len: '767',
    ref: 'swc-web.address.id',
    rows: 1n,
    filtered: 100,
    Extra: 'Using index'
  }
]

--- Running EXPLAIN ANALYZE for getTotalAdvocatesByState (CA) ---
[
  {
    EXPLAIN: '-> Group aggregate: count(`user`.id)  (cost=0.8 rows=1) (actual time=0.0156..0.0156 rows=0 loops=1)\n' +
      '    -> Nested loop inner join  (cost=0.7 rows=1) (actual time=0.0149..0.0149 rows=0 loops=1)\n' +
      "        -> Covering index lookup on address using address_administrative_area_level_1_country_code_idx (administrative_area_level_1='CA', country_code='US')  (cost=0.35 rows=1) (actual time=0.0138..0.0138 rows=0 loops=1)\n" +
      '        -> Covering index lookup on user using user_address_id_idx (address_id=address.id)  (cost=0.35 rows=1) (never executed)\n'
  }
]
--- Benchmark Summary (Parallel Execution) ---
Total states attempted: 55
Successful queries: 55
Total wall-clock time: 635.95ms (for all parallel queries)
Average time per successful state query: 590.80ms (individual query duration)
Min individual query time: 167.28ms
Max individual query time: 635.95ms
```
</details>

#### Summary
**Original Single-Query Approach:**
- **EXPLAIN Output**: The query planner uses the address_administrative_area_level_1_country_code_idx index. However, it's performing a range scan on address and then a ref lookup on user.
- **EXPLAIN ANALYZE Output**:  The planner estimates ~91k rows from the address table but the actual result is ~1.76 million rows. This significant underestimation can lead to poor query planning. The nested loop join is executed 1.76 million times.
- **Total Execution Time**: The single, large query consistently takes around **6.7 seconds**. 

**New Concurrent State-by-State Approach:**
- Total Execution Time: **~0.64 seconds** (for all 55 states in parallel). 
  - Performance Gain: This represents a ~10x improvement over the original query.
- Query Plan: For each state-specific query (like the 'CA' example), the database performs a much faster ref lookup (or "covering index lookup"). Because administrative_area_level_1 (state) is the leading column in the index, the database can instantly pinpoint the exact data for that state without scanning unnecessary rows.

Fixes: [PROD-SWC-WEB-4XM](https://stand-with-crypto.sentry.io/issues/5914493420/)